# Remove allocation in hashDigest to improve hashing speed