--- a/hashring_test.go
+++ b/hashring_test.go
@@ -476,3 +476,27 @@ func BenchmarkHashes(b *testing.B) {
 		hashRing.GetNodes(o.key, 2)
 	}
 }
+
+func BenchmarkHashesSingle(b *testing.B) {
+	nodes := []string{"a", "b", "c", "d", "e", "f", "g"}
+	hashRing := New(nodes)
+	tt := []struct {
+		key   string
+		nodes []string
+	}{
+		{"test", []string{"a", "b"}},
+		{"test", []string{"a", "b"}},
+		{"test1", []string{"b", "d"}},
+		{"test2", []string{"f", "b"}},
+		{"test3", []string{"f", "c"}},
+		{"test4", []string{"c", "b"}},
+		{"test5", []string{"f", "a"}},
+		{"aaaa", []string{"b", "a"}},
+		{"bbbb", []string{"f", "a"}},
+	}
+	b.ResetTimer()
+	for i := 0; i < b.N; i++ {
+		o := tt[i%len(tt)]
+		hashRing.GetNode(o.key)
+	}
+}