Using an constant-sized array rather than a slice lets Go allocate the hash
on the stack rather than the heap when possible.

```
benchmark                   old ns/op     new ns/op     delta
BenchmarkHashes-8           771           497           -35.54%
BenchmarkHashesSingle-8     492           239           -51.42%

benchmark                   old allocs     new allocs     delta
BenchmarkHashes-8           4              1              -75.00%
BenchmarkHashesSingle-8     3              0              -100.00%
```